prism-backend/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0                  # Production WSGI server (for Render/cloud deployment)

# Configuration & Environment
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# AI/ML Services
groq>=0.4.1
openai>=1.3.0                    # Optional OpenAI integration

# Databases & Storage
redis>=5.0.1
pinecone>=3.0.0
fastembed>=0.2.0
motor>=3.3.0                     # MongoDB async driver
neo4j>=5.14.0

# Security & Authentication
bcrypt>=4.1.0                    # Password hashing
python-jose[cryptography]>=3.3.0 # JWT tokens
passlib[bcrypt]>=1.7.4          # Password utilities
cryptography>=41.0.0            # Encryption utilities

# External Services
sendgrid>=6.10.0
duckduckgo-search>=3.9.0
requests>=2.31.0
httpx>=0.27.0
email-validator>=2.1.0

# Deep Research Agent - Browser Automation
playwright>=1.40.0
playwright-stealth>=1.0.6

# Utilities & Processing
python-multipart>=0.0.6         # File uploads
beautifulsoup4>=4.12.0          # Web scraping
youtube-search-python>=1.6.6    # YouTube search
edge-tts>=6.1.0                 # Text-to-speech
apscheduler>=3.10.0             # Task scheduling
pytz>=2023.0                    # Timezone handling
celery[redis]>=5.3.0            # Distributed task queue with Redis backend

# Development & Security Tools
python-decouple>=3.8            # Enhanced environment management
slowapi>=0.1.9                 # Rate limiting
python-multipart>=0.0.6        # File handling